+++ src/shared/utils/smp-gcpsid-raw.interface.ts
@@ -1,4 +1,0 @@
-export interface SmpSidRaw {
-    expiry: number;
-    value: string;
-}